pyproject.toml
[project]
name = "fastapi-django-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Stepbus",email = "aleksovs99@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "django-timezone-field (>=7.1,<8.0)",
    "django-celery-beat (>=2.7.0,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
